removed session helper
