+++ task_board/app/controllers/application_controller.rb
@@ -1,9 +1,9 @@
 class ApplicationController < ActionController::Base
-  include SessionsHelper
-
   rescue_from Exception, with: :render_500
   rescue_from ActiveRecord::RecordNotFound, with: :render_404
   rescue_from ActionController::RoutingError, with: :render_404
+
+  helper_method :current_user, :logged_in?
 
   def routing_error
     raise ActionController::RoutingError, params[:path]
@@ -18,6 +18,23 @@ class ApplicationController < ActionController::Base
   end
 
   private
+
+  def log_in(user)
+    session[:user_id] = user.id
+  end
+
+  def current_user
+    @current_user ||= User.find_by(id: session[:user_id]) if session[:user_id]
+  end
+
+  def logged_in?
+    current_user
+  end
+
+  def log_out
+    session.delete(:user_id)
+    @current_user = nil
+  end
 
   def logged_in_user
     redirect_to login_url unless logged_in?
