--- task_board/app/helpers/sessions_helper.rb
@@ -1,18 +1,0 @@
-module SessionsHelper
-  def log_in(user)
-    session[:user_id] = user.id
-  end
-
-  def current_user
-    @current_user ||= User.find_by(id: session[:user_id]) if session[:user_id]
-  end
-
-  def logged_in?
-    current_user
-  end
-
-  def log_out
-    session.delete(:user_id)
-    @current_user = nil
-  end
-end
